added basic pixel-wise classification and testing
Haven't run the ClassifyTest FSM on the robot yet, but I think it should
work (will do that tomorrow morning). Basic idea is we can just swap out
different algorithms in a line or two, and see how they work (and do the
same thing in our final occupancy-grid-building code).