fixed SpiralScan typo, some updates to grid
Some minor changes to 'grid.py' which I don't quite remember,
but were made before the demo (maybe just grid size increase).
Fixed a slight typo in the new SpiralScan, so it now runs to
completion and saves a bunch of patches. Not sure if they're
the right patches yet, but probably.